requirements.txt
python-binance
supabase
httpx
python-dateutil
python-dotenv
ccxt>=4.0.0